% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_network_metadata.R
\name{campbellMetaUpdate}
\alias{campbellMetaUpdate}
\title{Update CAMPBELL metadata.}
\usage{
campbellMetaUpdate()
}
\description{
Update metadata for CAMPBELL network and upload to ADT database.
}
