% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_network_metadata.R
\name{microstepMetaUpdate}
\alias{microstepMetaUpdate}
\title{Update MICROSTEP metadata.}
\usage{
microstepMetaUpdate()
}
\description{
Update metadata for MICROSTEP network and upload to ADT database.
}
