% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{process.seba}
\alias{process.seba}
\title{Process SEBA data.}
\usage{
process.seba(dirFTP, dirAWS, dirUP = NULL, upload = TRUE)
}
\arguments{
\item{dirFTP}{full path to the directory containing the Seba data.\cr}

\item{dirAWS}{full path to the directory containing the AWS_DATA folder on Seba server.\cr}

\item{dirUP}{full path to the directory containing the AWS_DATA folder on ADT server.
Default NULL, must be provided if \code{upload} is \code{TRUE}.\cr}

\item{upload}{logical, if TRUE the data will be uploaded to ADT server.}
}
\description{
Get the data from SEBA database, parse and insert into ADT database.
}
